refactor: split the yt-dlp updater at its own seams
